financial_health_score_agent/pyproject.toml
[project]
name = "financial-health-score-agent"
version = "0.1.0"
description = "AI-driven agent that calculates comprehensive Financial Health Scores (0-1000) with trend analysis and actionable recommendations"
authors = [{ name = "Financial Health Team", email = "team@example.com" }]
license = "Apache License 2.0"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "google-adk==1.3.0",
    "google-cloud-aiplatform>=1.93.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "black>=25.1.0",
    "pytest-asyncio>=0.26.0",
    "jupyter>=1.0.0",
]

deployment = [
    "absl-py>=2.2.1",
    "uvicorn>=0.20.0",
    "fastapi>=0.100.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api" 